+++ common/src/main/java/dev/matthiesen/common/cobblemon_boosters/gui/gooey/screens/subscreens/BucketBoostBuilderGui.java
@@ -8,6 +8,7 @@ import ca.landonjw.gooeylibs2.api.page.Page;
 import ca.landonjw.gooeylibs2.api.template.types.ChestTemplate;
 import dev.matthiesen.common.cobblemon_boosters.CobblemonBoosters;
 import dev.matthiesen.common.cobblemon_boosters.data.SpawnBucketBoost;
+import dev.matthiesen.common.cobblemon_boosters.gui.gooey.screens.utils.Helpers;
 import dev.matthiesen.common.cobblemon_boosters.utils.MenuUtils;
 import dev.matthiesen.common.cobblemon_boosters.utils.TextUtils;
 import net.minecraft.network.chat.Component;
@@ -25,8 +26,8 @@ public class BucketBoostBuilderGui {
     public String currentMode = null;
     public final Consumer<SpawnBucketBoost> setActiveBoost;
 
-    public final List<String> allowedBuckets = List.of("common", "uncommon", "rare", "ultra-rare");
-    public final List<String> allowedUnits = List.of("seconds", "minutes", "hours", "days");
+    public final List<String> allowedBuckets = Helpers.allowedBuckets;
+    public final List<String> allowedUnits = Helpers.allowedUnits;
     public final Map<String, String> labelToColor = Map.of(
             "Multiplier", "<green>",
             "Duration", "<aqua>",
@@ -65,7 +66,7 @@ public class BucketBoostBuilderGui {
         }
 
         public BoostBuilder setUnit(String unit) {
-            List<String> allowedUnits = List.of("seconds", "minutes", "hours", "days");
+            List<String> allowedUnits = Helpers.allowedUnits;
             if (allowedUnits.contains(unit.toLowerCase())) {
                 this.unit = unit.toLowerCase();
             } else {
@@ -74,22 +75,11 @@ public class BucketBoostBuilderGui {
             return this;
         }
 
-        public static int parseTotalSeconds(int duration, String unit) {
-            int totalSeconds;
-            switch (unit) {
-                case "minutes" -> totalSeconds = duration * 60;
-                case "hours" -> totalSeconds = duration * 3600;
-                case "days" -> totalSeconds = duration * 86400;
-                default -> totalSeconds = duration;
-            }
-            return totalSeconds;
-        }
-
         public SpawnBucketBoost build() {
             try {
                 SpawnBucketBoost boost = new SpawnBucketBoost();
                 boost.setMultiplier(multiplier);
-                int totalSeconds = parseTotalSeconds(duration, unit);
+                int totalSeconds = Helpers.parseTotalSeconds(duration, unit);
                 boost.setDuration(totalSeconds);
                 boost.setBucket(bucket);
                 return boost;
@@ -134,12 +124,6 @@ public class BucketBoostBuilderGui {
         return TextUtils.deserializeMC(
                 TextUtils.parse("Cobblemon Boosters - Boost Builder")
         );
-    }
-
-    public Button getFrame() {
-        return GooeyButton.builder()
-                .display(MenuUtils.getFrameItem())
-                .build();
     }
 
     public Button buildModeButton(
@@ -398,7 +382,7 @@ public class BucketBoostBuilderGui {
         if (isReadyToConfirm())
             builder = builder.set(1, 7, getConfirmButton());
 
-        return builder.fill(getFrame()).build();
+        return builder.fill(Helpers.getFrame()).build();
     }
 
     public Page getPage() {
